+++ example/lib/example5_themes_playground/theme/flex_theme_light.dart
@@ -264,6 +264,46 @@ FlexColorScheme flexColorSchemeLight(ThemeController controller, Color source) {
             sliderTrackHeight: controller.sliderTrackHeight,
             sliderYear2023: controller.sliderYear2023,
             progressIndicatorYear2023: controller.progressIndicatorYear2023,
+            progressIndicatorBaseSchemeColor:
+                controller.progressIndicatorBaseSchemeColor,
+            progressIndicatorLinearTrackSchemeColor:
+                controller.progressIndicatorLinearTrackSchemeColor,
+            progressIndicatorCircularTrackSchemeColor:
+                controller.progressIndicatorCircularTrackSchemeColor,
+            progressIndicatorRefreshBackgroundSchemeColor:
+                controller.progressIndicatorRefreshBackgroundSchemeColor,
+            progressIndicatorStopIndicatorSchemeColor:
+                controller.progressIndicatorStopIndicatorSchemeColor,
+            progressIndicatorLinearMinHeight:
+                controller.progressIndicatorLinearMinHeight,
+            progressIndicatorLinearRadius:
+                controller.progressIndicatorLinearRadius,
+            progressIndicatorStopIndicatorRadius:
+                controller.progressIndicatorStopIndicatorRadius,
+            progressIndicatorStrokeWidth: controller.progressIndicatorStrokeWidth,
+            progressIndicatorStrokeAlign:
+                controller.progressIndicatorStrokeAlign,
+            progressIndicatorTrackGap: controller.progressIndicatorTrackGap,
+            progressIndicatorStrokeCap: controller.progressIndicatorStrokeCap,
+            progressIndicatorConstraints:
+                controller.progressIndicatorConstraints,
+            progressIndicatorCircularTrackPadding:
+                controller.progressIndicatorCircularTrackPaddingStart != null ||
+                        controller.progressIndicatorCircularTrackPaddingEnd !=
+                            null ||
+                        controller.progressIndicatorCircularTrackPaddingTop !=
+                            null ||
+                        controller.progressIndicatorCircularTrackPaddingBottom !=
+                            null
+                    ? EdgeInsetsDirectional.fromSTEB(
+                        controller.progressIndicatorCircularTrackPaddingStart ??
+                            0,
+                        controller.progressIndicatorCircularTrackPaddingTop ?? 0,
+                        controller.progressIndicatorCircularTrackPaddingEnd ?? 0,
+                        controller.progressIndicatorCircularTrackPaddingBottom ??
+                            0,
+                      )
+                    : null,
             // Input decoration (TextField) settings.
             // Base ColorScheme used by TextField InputDecorator.
             inputDecoratorSchemeColor: