+++ example/lib/example5_themes_playground/widgets/panels/progress_indicator/progress_indicator_panel.dart
@@ -3,8 +3,13 @@ import 'package:flutter/material.dart';
 import '../../../../shared/controllers/theme_controller.dart';
 import '../../../../shared/utils/link_text_span.dart';
 import '../../../../shared/widgets/universal/list_tile_reveal.dart';
+import '../../../../shared/widgets/universal/responsive_two_widgets.dart';
 import '../../../../shared/widgets/universal/showcase_material.dart';
+import '../../../../shared/widgets/universal/slider_list_tile_reveal.dart';
 import '../../../../shared/widgets/universal/switch_list_tile_reveal.dart';
+import '../../../theme/theme_values.dart';
+import '../../shared/color_scheme_popup_menu.dart';
+import '../../shared/enum_popup_menu.dart';
 import '../../shared/test_adaptive_response.dart';
 
 class ProgressIndicatorPanel extends StatelessWidget {
@@ -29,11 +34,33 @@ class ProgressIndicatorPanel extends StatelessWidget {
       color: theme.colorScheme.primary,
       fontWeight: FontWeight.bold,
     );
-    // final bool isLight = theme.brightness == Brightness.light;
 
     // The most common logic for enabling Playground controls.
     final bool enableControl =
         controller.useSubThemes && controller.useFlexColorScheme;
+
+    // Default labels for SchemeColor properties
+    final String labelBaseDefault =
+        controller.progressIndicatorBaseSchemeColor == null
+            ? 'primary'
+            : '${controller.progressIndicatorBaseSchemeColor?.name}';
+    final String labelLinearTrackDefault =
+        controller.progressIndicatorLinearTrackSchemeColor == null
+            ? 'surfaceVariant'
+            : '${controller.progressIndicatorLinearTrackSchemeColor?.name}';
+    final String labelCircularTrackDefault =
+        controller.progressIndicatorCircularTrackSchemeColor == null
+            ? 'surfaceVariant'
+            : '${controller.progressIndicatorCircularTrackSchemeColor?.name}';
+    final String labelRefreshBackgroundDefault = controller
+                .progressIndicatorRefreshBackgroundSchemeColor ==
+            null
+        ? 'surface'
+        : '${controller.progressIndicatorRefreshBackgroundSchemeColor?.name}';
+    final String labelStopIndicatorDefault =
+        controller.progressIndicatorStopIndicatorSchemeColor == null
+            ? 'primary'
+            : '${controller.progressIndicatorStopIndicatorSchemeColor?.name}';
 
     return Column(
       crossAxisAlignment: CrossAxisAlignment.start,
@@ -57,6 +84,221 @@ class ProgressIndicatorPanel extends StatelessWidget {
           value: use2023Style,
           onChanged: controller.setProgressIndicatorYear2023,
         ),
+        ResponsiveTwoWidgets(builder: (BuildContext context, bool isRow) {
+          return RowOrColumn(
+            firstWidget: ColorSchemePopupMenu(
+              contentPadding: ThemeValues.tilePaddingStart(context, isRow),
+              enabled: enableControl,
+              title: const Text('Base color'),
+              defaultLabel: labelBaseDefault,
+              value: controller.progressIndicatorBaseSchemeColor,
+              onChanged: controller.setProgressIndicatorBaseSchemeColor,
+            ),
+            lastWidget: ColorSchemePopupMenu(
+              contentPadding: ThemeValues.tilePaddingEnd(context, isRow),
+              enabled: enableControl,
+              title: const Text('Linear track color'),
+              defaultLabel: labelLinearTrackDefault,
+              value: controller.progressIndicatorLinearTrackSchemeColor,
+              onChanged: controller.setProgressIndicatorLinearTrackSchemeColor,
+            ),
+            isRow: isRow,
+          );
+        }),
+        ResponsiveTwoWidgets(builder: (BuildContext context, bool isRow) {
+          return RowOrColumn(
+            firstWidget: ColorSchemePopupMenu(
+              contentPadding: ThemeValues.tilePaddingStart(context, isRow),
+              enabled: enableControl,
+              title: const Text('Circular track color'),
+              defaultLabel: labelCircularTrackDefault,
+              value: controller.progressIndicatorCircularTrackSchemeColor,
+              onChanged:
+                  controller.setProgressIndicatorCircularTrackSchemeColor,
+            ),
+            lastWidget: ColorSchemePopupMenu(
+              contentPadding: ThemeValues.tilePaddingEnd(context, isRow),
+              enabled: enableControl,
+              title: const Text('Refresh background color'),
+              defaultLabel: labelRefreshBackgroundDefault,
+              value: controller.progressIndicatorRefreshBackgroundSchemeColor,
+              onChanged:
+                  controller.setProgressIndicatorRefreshBackgroundSchemeColor,
+            ),
+            isRow: isRow,
+          );
+        }),
+        ResponsiveTwoWidgets(builder: (BuildContext context, bool isRow) {
+          return RowOrColumn(
+            firstWidget: ColorSchemePopupMenu(
+              contentPadding: ThemeValues.tilePaddingStart(context, isRow),
+              enabled: enableControl,
+              title: const Text('Stop indicator color'),
+              defaultLabel: labelStopIndicatorDefault,
+              value: controller.progressIndicatorStopIndicatorSchemeColor,
+              onChanged:
+                  controller.setProgressIndicatorStopIndicatorSchemeColor,
+            ),
+            lastWidget: EnumPopupMenu<StrokeCap>(
+              contentPadding: ThemeValues.tilePaddingEnd(context, isRow),
+              enabled: enableControl,
+              values: StrokeCap.values,
+              title: const Text('Stroke cap'),
+              value: controller.progressIndicatorStrokeCap,
+              onChanged: controller.setProgressIndicatorStrokeCap,
+            ),
+            isRow: isRow,
+          );
+        }),
+        SliderListTileReveal(
+          enabled: enableControl,
+          title: const Text('Linear minimum height'),
+          value: controller.progressIndicatorLinearMinHeight,
+          onChanged: controller.setProgressIndicatorLinearMinHeight,
+          min: 1,
+          max: 20,
+          divisions: 19,
+          valueDecimalPlaces: 0,
+          valueHeading: 'HEIGHT',
+          valueUnitLabel: ' dp',
+          valueDefaultLabel: '4 dp',
+        ),
+        SliderListTileReveal(
+          enabled: enableControl,
+          title: const Text('Linear radius'),
+          value: controller.progressIndicatorLinearRadius,
+          onChanged: controller.setProgressIndicatorLinearRadius,
+          min: 0,
+          max: 20,
+          divisions: 40,
+          valueDecimalPlaces: 1,
+          valueHeading: 'RADIUS',
+          valueUnitLabel: ' dp',
+          valueDefaultLabel: '0 dp',
+        ),
+        SliderListTileReveal(
+          enabled: enableControl,
+          title: const Text('Stop indicator radius'),
+          value: controller.progressIndicatorStopIndicatorRadius,
+          onChanged: controller.setProgressIndicatorStopIndicatorRadius,
+          min: 0,
+          max: 20,
+          divisions: 40,
+          valueDecimalPlaces: 1,
+          valueHeading: 'RADIUS',
+          valueUnitLabel: ' dp',
+          valueDefaultLabel: '0 dp',
+        ),
+        SliderListTileReveal(
+          enabled: enableControl,
+          title: const Text('Stroke width'),
+          value: controller.progressIndicatorStrokeWidth,
+          onChanged: controller.setProgressIndicatorStrokeWidth,
+          min: 1,
+          max: 10,
+          divisions: 18,
+          valueDecimalPlaces: 1,
+          valueHeading: 'WIDTH',
+          valueUnitLabel: ' dp',
+          valueDefaultLabel: '4 dp',
+        ),
+        SliderListTileReveal(
+          enabled: enableControl,
+          title: const Text('Stroke align'),
+          value: controller.progressIndicatorStrokeAlign,
+          onChanged: controller.setProgressIndicatorStrokeAlign,
+          min: -1,
+          max: 1,
+          divisions: 40,
+          valueDecimalPlaces: 2,
+          valueHeading: 'ALIGN',
+          valueUnitLabel: '',
+          valueDefaultLabel: '0',
+        ),
+        SliderListTileReveal(
+          enabled: enableControl,
+          title: const Text('Track gap'),
+          value: controller.progressIndicatorTrackGap,
+          onChanged: controller.setProgressIndicatorTrackGap,
+          min: 0,
+          max: 10,
+          divisions: 20,
+          valueDecimalPlaces: 1,
+          valueHeading: 'GAP',
+          valueUnitLabel: ' dp',
+          valueDefaultLabel: '0 dp',
+        ),
+        const Divider(),
+        const ListTile(title: Text('Circular track padding')),
+        ResponsiveTwoWidgets(builder: (BuildContext context, bool isRow) {
+          return RowOrColumn(
+            firstWidget: SliderListTileReveal(
+              contentPadding: ThemeValues.tilePaddingStart(context, isRow),
+              enabled: enableControl,
+              title: const Text('Padding start'),
+              value: controller.progressIndicatorCircularTrackPaddingStart,
+              onChanged:
+                  controller.setProgressIndicatorCircularTrackPaddingStart,
+              min: 0,
+              max: 20,
+              divisions: 40,
+              valueDecimalPlaces: 1,
+              valueHeading: 'START',
+              valueUnitLabel: ' dp',
+              valueDefaultLabel: '0 dp',
+            ),
+            lastWidget: SliderListTileReveal(
+              contentPadding: ThemeValues.tilePaddingEnd(context, isRow),
+              enabled: enableControl,
+              title: const Text('Padding end'),
+              value: controller.progressIndicatorCircularTrackPaddingEnd,
+              onChanged: controller.setProgressIndicatorCircularTrackPaddingEnd,
+              min: 0,
+              max: 20,
+              divisions: 40,
+              valueDecimalPlaces: 1,
+              valueHeading: 'END',
+              valueUnitLabel: ' dp',
+              valueDefaultLabel: '0 dp',
+            ),
+            isRow: isRow,
+          );
+        }),
+        ResponsiveTwoWidgets(builder: (BuildContext context, bool isRow) {
+          return RowOrColumn(
+            firstWidget: SliderListTileReveal(
+              contentPadding: ThemeValues.tilePaddingStart(context, isRow),
+              enabled: enableControl,
+              title: const Text('Padding top'),
+              value: controller.progressIndicatorCircularTrackPaddingTop,
+              onChanged: controller.setProgressIndicatorCircularTrackPaddingTop,
+              min: 0,
+              max: 20,
+              divisions: 40,
+              valueDecimalPlaces: 1,
+              valueHeading: 'TOP',
+              valueUnitLabel: ' dp',
+              valueDefaultLabel: '0 dp',
+            ),
+            lastWidget: SliderListTileReveal(
+              contentPadding: ThemeValues.tilePaddingEnd(context, isRow),
+              enabled: enableControl,
+              title: const Text('Padding bottom'),
+              value: controller.progressIndicatorCircularTrackPaddingBottom,
+              onChanged:
+                  controller.setProgressIndicatorCircularTrackPaddingBottom,
+              min: 0,
+              max: 20,
+              divisions: 40,
+              valueDecimalPlaces: 1,
+              valueHeading: 'BOTTOM',
+              valueUnitLabel: ' dp',
+              valueDefaultLabel: '0 dp',
+            ),
+            isRow: isRow,
+          );
+        }),
+        const Divider(),
         const Padding(
           padding: EdgeInsets.all(16.0),
           child: ProgressIndicatorShowcase(),